Login Page and Attached Properties Video 6
A login page was created that allows a tag to be display on the field before the user types anything. Once the user types, the tag desappears and the actual text is display. The base PasswordBox behavior was modified to behave the same with the tag. To modified the PasswordBox, attached properties were created.